# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ledger API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/transactions/deposit": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Deposit Endpoint",
        "operationId": "create_deposit_endpoint_transactions_deposit_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}/reverse": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Reversal Endpoint",
        "operationId": "create_reversal_endpoint_transactions__transaction_id__reverse_post",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/transfer": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Transfer Endpoint",
        "operationId": "create_transfer_endpoint_transactions_transfer_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DepositCreate": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "cash_account_id": {
            "type": "string",
            "title": "Cash Account Id"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "cash_account_id",
          "amount_minor"
        ],
        "title": "DepositCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostingResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "side",
          "amount_minor"
        ],
        "title": "PostingResponse"
      },
...
      "TransactionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "postings": {
            "items": {
              "$ref": "#/components/schemas/PostingResponse"
            },
            "type": "array",
            "title": "Postings"
          },
          "reversal_of_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reversal Of Id"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "type",
          "state",
          "postings"
        ],
        "title": "TransactionResponse"
      },
      "TransferCreate": {
        "properties": {
          "from_account_id": {
            "type": "string",
            "title": "From Account Id"
          },
          "to_account_id": {
            "type": "string",
            "title": "To Account Id"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          }
        },
        "type": "object",
        "required": [
          "from_account_id",
          "to_account_id",
          "amount_minor"
        ],
        "title": "TransferCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}